161 capstone modified
Create a limited database that assigns computer games to a variety of consoles, information regarding the games includes title, genre, developer, year. The code should allow the user to add/remove a game from a console's dictionary, and remove a console from the list of consoles. Returns where a user can play a game.